Fix vari

+++ scripts/3.app.sh
@@ -1,29 +1,36 @@
 #!/usr/bin/env bash
 
 # Caricamento file env.conf
-#source ${HOME}/startup/env.conf
+source ${HOME}/startup/env.conf
 
 waitForInput() {
     printf "Premere un tasto per cleanup e riavvio..."
     read -n 1 -s
 }
 
+saveLog() {
+    messageToLog="$1"
+    printf "%s\n" "$messageToLog" | tee -a "${INSTALL_LOG}"
+}
+
+saveLogAndExit() {
+    saveLog "$1"
+    exit
+}
+
 getEthName() {
     ETHNAME="eu-lan"    
-    #echo f3rt3c | sudo -S nmcli connection modify Wired\ connection\ 1 con-name ${ETHNAME}
+    echo ${PASSWORD} | sudo -S nmcli connection modify Wired\ connection\ 1 con-name ${ETHNAME}
 }
 
 setIpAddress() {
     if [ "${PROJECT_NAME,,}" = "neuron" ]; then
         getEthName
-        if [ -z ${ETHNAME} ]; then 
-            echo "ERROR! No active ethernet interface found"; 
-            exit
-        fi
+        if [ -z ${ETHNAME} ]; then saveLogAndExit "ERROR! No active ethernet interface found"; fi
 
-        echo f3rt3c | sudo -S nmcli connection modify ${ETHNAME} ipv4.addresses 192.168.10.111/16
-        echo f3rt3c | sudo -S nmcli connection modify ${ETHNAME} ipv4.gateway 192.168.10.254
-        echo f3rt3c | sudo -S nmcli connection up ${ETHNAME}
+        echo ${PASSWORD} | sudo -S nmcli connection modify ${ETHNAME} ipv4.addresses ${IP}/24
+        echo ${PASSWORD} | sudo -S nmcli connection modify ${ETHNAME} ipv4.gateway ${GTW}
+        echo ${PASSWORD} | sudo -S nmcli connection up ${ETHNAME}
     fi
 }
 
@@ -32,11 +39,12 @@ cleanupAndReboot() {
     sed -i '/source /home/fertec/startup/3\.app\.sh/d' ${HOME}/.bashrc
 
     # rimozione della cartella contenente gli script di avvio
-    rm -r ${HOME}/startup
+    if [ -d ${HOME}/startup ]; then rm -r ${HOME}/startup; fi
 
-    #sudo reboot
+    # riavvio
+    echo ${PASSWORD} | sudo -S reboot
 }
-PROJECT_NAME="neuron"
+
 setIpAddress
 
 #####################################
@@ -44,5 +52,5 @@ setIpAddress
 #####################################
 
 # Pulizia e riavvio
-#waitForInput
-#cleanupAndReboot
+waitForInput
+cleanupAndReboot
